+++ src/app/[pages]/layout.tsx
@@ -1,14 +1,4 @@
 import { Flex } from '@/components/style/Flex'
-// import './globals.css'
-import type { Metadata } from 'next'
-// import {Pretendard} from 'next/font'
-import { Noto_Sans_KR } from 'next/font/google'
-import Image from 'next/image';
-import path from '@/values/asset/path';
-import profile from '#/profile/JYP.png'
-// import { pretendard } from "../../node_modules/pretendard/dist/web/static/woff.css";
-
-// const noto = Noto_Sans_KR({ subsets: ['latin'] });
 
 export default function PagesLayout({
     children,
