+++ src/app/page.tsx
@@ -1,9 +1,4 @@
 import { Flex } from "@/components/style/Flex";
-import Temp from "@/components/style/Temp";
-
-import { ReactNode } from "react";
-import { db } from "./layout";
-import { collection,getDocs } from "firebase/firestore";
 
 type TempData = {
   name: string,
@@ -11,9 +6,6 @@ type TempData = {
 }
 
 export default async function MainSection() {
-  const querySnapshot = await getDocs(collection(db, "Routers"));
-  // const temp = querySnapshot.forEach(item => console.log(item));
-  // console.log(temp);
   return (
     <>
       {'블로그 메인'}
